first merge from jared into devops branch

+++ social-media-angular/src/app/components/edit-profile/edit-profile.component.ts
@@ -24,11 +24,13 @@ export class EditProfileComponent implements OnInit {
   fileName = '';
   hide = true;
   userId: number;
+  
 
   
   constructor(private router: Router, private _userService: UserService, private _authService: AuthService, private cookieService: CookieService) { 
     this.userService = _userService;
     this.authService = _authService;
+    this.cookieService = cookieService;
   }
 
   // user: User = {} as User;
@@ -51,60 +53,26 @@ export class EditProfileComponent implements OnInit {
   }
 
   UpdateUser() {
-    this.user.userId = this.authService.currentUser.userId;
+    //this.user.userId = this.authService.currentUser.userId;
+
     console.log(this.user.userId)
     this.userService.UpdateUser(this.user).subscribe(updateUser => {
       console.log(updateUser);
+      alert('Successfuly Updated User!');
     })
   }
 
   deleteAccount() {
-    this.user.userId = this.authService.currentUser.userId;
-    console.log("user to be deleted: " + this.user.userId)
-    this.userService.DeleteUser(this.user.userId!).subscribe()
-    alert('Successfully Deleted Account');
-    this.router.navigate(["login"])
-  }
+    //this.user.userId = this.authService.currentUser.userId;
 
-  profilePicture: string;
-  profilePictureDialog() {
-    const dialogRef = this.dialog.open(UploadProfilePictureDialog, {
-      width: '250px',
-      data: { profilePicutre: this.profilePicture},
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog is closed');
-      this.profilePicture = result;
-    });
-  }
-
-  onFileSelected(event: any) {
-
-    const file:File = event.target.files[0];
-
-    if (file) {
-        this.fileName = file.name;
-        const formData = new FormData();
-        formData.append("thumbnail", file);
-        const upload$ = this.userService.UploadImage(formData);
-        //upload$.subscribe();
+    let text = "Sorry to see you go, confirm to delete your account.";
+    if (confirm(text) == true) {
+      console.log("user to be deleted: " + this.user.userId)
+      this.userService.DeleteUser(this.user.userId!).subscribe()
+      alert('Successfully Deleted Account');
+      this.router.navigate(["login"])
+    } else {
+      this.router.navigate(["edit-profile"]);
     }
-  }
-
-}
-
-@Component({
-  selector: './upload-profile-picture-dialog',
-  templateUrl: './upload-profile-picture-dialog.html',
-})
-export class UploadProfilePictureDialog {
-  constructor(
-    public dialogRef: MatDialogRef<UploadProfilePictureDialog>,
-    @Inject(MAT_DIALOG_DATA) public data: User,
-  ) {}
-
-  onNoClick(): void {
-    this.dialogRef.close();
   }
 }
